drivers: tt_fuse: move functional fuse and efuse code into a driver
Move functional fuse and efuse code into a driver, out of the bh_arc
library
