+++ app/smc/prj.conf
@@ -38,3 +38,6 @@ CONFIG_TT_FW_TABLE=y
 # Enable required drivers
 CONFIG_TT_TIMER=y
 CONFIG_TT_BH_ARC_TIMER=y
+
+CONFIG_TT_FUSE=y
+CONFIG_TT_BH_ARC_EFUSE=y
