+++ include/tenstorrent/drivers/fuse/efuse.h
@@ -8,8 +8,6 @@
 
 #include <stdint.h>
 
-#include <arc_dma.h>
-
 /* Useful addresses */
 #define FUSE_ASIC_ID_ADDR 50
