+++ lib/tenstorrent/bh_arc/eth.c
@@ -11,7 +11,7 @@
 #include <tenstorrent/reg.h>
 #include <tenstorrent/fw_table/read_only_table.h>
 #include <tenstorrent/fw_table/fw_table.h>
-#include "efuse.h"
+#include <tenstorrent/drivers/fuse/efuse.h>
 
 #define ETH_SETUP_TLB  0
 #define ETH_PARAM_ADDR 0x7c000
