+++ lib/tenstorrent/bh_arc/harvesting.c
@@ -7,7 +7,7 @@
 #include <zephyr/sys/util.h>
 #include "harvesting.h"
 #include <tenstorrent/fw_table/fw_table.h>
-#include "functional_efuse.h"
+#include <tenstorrent/drivers/fuse/functional_efuse.h>
 #include "noc.h"
 
 TileEnable tile_enable = {
